+++ tests/acceptance/test_decorators.py
@@ -3,9 +3,9 @@ import random
 from unittest import mock
 
 import pytest
-from tests.utils import Keys
 
 from aiocache import cached, cached_stampede, multi_cached
+from ..utils import Keys
 
 
 async def return_dict(keys=None):
@@ -15,21 +15,18 @@ async def return_dict(keys=None):
     return ret
 
 
-async def stub(*args, key=None, seconds=0, **kwargs):
+async def stub(arg: float, seconds: int = 0) -> str:
     await asyncio.sleep(seconds)
-    if key:
-        return str(key)
     return str(random.randint(1, 50))
 
 
 class TestCached:
     @pytest.fixture(autouse=True)
     def default_cache(self, mocker, cache):
-        mocker.patch("aiocache.decorators._get_cache", return_value=cache)
+        mocker.patch("aiocache.decorators._get_cache", autospec=True, return_value=cache)
 
-    @pytest.mark.asyncio
     async def test_cached_ttl(self, cache):
-        @cached(ttl=1, key=Keys.KEY)
+        @cached(ttl=2, key=Keys.KEY)
         async def fn():
             return str(random.randint(1, 50))
 
@@ -37,10 +34,9 @@ class TestCached:
         resp2 = await fn()
 
         assert await cache.get(Keys.KEY) == resp1 == resp2
-        await asyncio.sleep(1)
+        await asyncio.sleep(2.1)
         assert await cache.get(Keys.KEY) is None
 
-    @pytest.mark.asyncio
     async def test_cached_key_builder(self, cache):
         def build_key(f, self, a, b):
             return "{}_{}_{}_{}".format(self, f.__name__, a, b)
@@ -56,23 +52,21 @@ class TestCached:
 class TestCachedStampede:
     @pytest.fixture(autouse=True)
     def default_cache(self, mocker, cache):
-        mocker.patch("aiocache.decorators._get_cache", return_value=cache)
+        mocker.patch("aiocache.decorators._get_cache", autospec=True, return_value=cache)
 
-    @pytest.mark.asyncio
-    @pytest.mark.skip(reason="flaky test")
     async def test_cached_stampede(self, mocker, cache):
         mocker.spy(cache, "get")
         mocker.spy(cache, "set")
-        decorator = cached_stampede(ttl=10, lease=2)
+        decorator = cached_stampede(ttl=10, lease=3)
 
         await asyncio.gather(decorator(stub)(0.5), decorator(stub)(0.5))
 
-        cache.get.assert_called_with("acceptance.test_decoratorsstub(0.5,)[]")
+        cache.get.assert_called_with("tests.acceptance.test_decoratorsstub(0.5,)[]")
         assert cache.get.call_count == 4
-        cache.set.assert_called_with("acceptance.test_decoratorsstub(0.5,)[]", mock.ANY, ttl=10)
-        assert cache.set.call_count == 1
+        cache.set.assert_called_with("tests.acceptance.test_decoratorsstub(0.5,)[]",
+                                     mock.ANY, ttl=10)
+        assert cache.set.call_count == 1, cache.set.call_args_list
 
-    @pytest.mark.asyncio
     async def test_locking_dogpile_lease_expiration(self, mocker, cache):
         mocker.spy(cache, "get")
         mocker.spy(cache, "set")
@@ -87,8 +81,7 @@ class TestCachedStampede:
         assert cache.get.call_count == 6
         assert cache.set.call_count == 3
 
-    @pytest.mark.asyncio
-    async def test_locking_dogpile_task_cancellation(self, mocker, cache):
+    async def test_locking_dogpile_task_cancellation(self, cache):
         @cached_stampede()
         async def cancel_task():
             raise asyncio.CancelledError()
@@ -100,9 +93,8 @@ class TestCachedStampede:
 class TestMultiCachedDecorator:
     @pytest.fixture(autouse=True)
     def default_cache(self, mocker, cache):
-        mocker.patch("aiocache.decorators._get_cache", return_value=cache)
+        mocker.patch("aiocache.decorators._get_cache", autospec=True, return_value=cache)
 
-    @pytest.mark.asyncio
     async def test_multi_cached(self, cache):
         multi_cached_decorator = multi_cached("keys")
 
@@ -112,7 +104,6 @@ class TestMultiCachedDecorator:
         for key in default_keys:
             assert await cache.get(key) is not None
 
-    @pytest.mark.asyncio
     async def test_keys_without_kwarg(self, cache):
         @multi_cached("keys")
         async def fn(keys):
@@ -121,20 +112,19 @@ class TestMultiCachedDecorator:
         await fn([Keys.KEY])
         assert await cache.exists(Keys.KEY) is True
 
-    @pytest.mark.asyncio
     async def test_multi_cached_key_builder(self, cache):
+        # TODO(PY311): Remove str() calls
         def build_key(key, f, self, keys, market="ES"):
-            return "{}_{}_{}".format(f.__name__, key, market)
+            return "{}_{}_{}".format(f.__name__, str(key), market)
 
         @multi_cached(keys_from_attr="keys", key_builder=build_key)
         async def fn(self, keys, market="ES"):
             return {Keys.KEY: 1, Keys.KEY_1: 2}
 
         await fn("self", keys=[Keys.KEY, Keys.KEY_1])
-        assert await cache.exists("fn_" + Keys.KEY + "_ES") is True
-        assert await cache.exists("fn_" + Keys.KEY_1 + "_ES") is True
+        assert await cache.exists("fn_" + str(Keys.KEY) + "_ES") is True
+        assert await cache.exists("fn_" + str(Keys.KEY_1) + "_ES") is True
 
-    @pytest.mark.asyncio
     async def test_fn_with_args(self, cache):
         @multi_cached("keys")
         async def fn(keys, *args):
@@ -144,7 +134,6 @@ class TestMultiCachedDecorator:
         await fn([Keys.KEY], "arg")
         assert await cache.exists(Keys.KEY) is True
 
-    @pytest.mark.asyncio
     async def test_double_decorator(self, cache):
         def dummy_d(fn):
             async def wrapper(*args, **kwargs):