+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
 
+## 0.12.0 (2023-xx-xx)
+
+* Add ``async with`` support to ``BaseCache``.
+* Remove deprecated ``loop`` parameters.
+* Remove deprecated ``cache`` parameter from ``create()``.
+* Use ``str()`` in ``_build_key()`` to ensure consistency of enum keys between different Python versions (if using enum keys, upgrading to 0.12 may invalidate existing caches due to key values changing).
+
 ## 0.11.1 (2019-07-31)
 
 * Don't hardcode import redis and memcached in factory [#461](https://github.com/argaen/aiocache/issues/461) - Manuel Miranda